SB-112 (refactor) : 각 모듈 user pakage 내 intercetpor, resolver pagkage 이동
SB-112 (refactor) : 각 모듈 user pakage 내 intercetpor, resolver pagkage 이동
