+++ account/src/main/java/account/domain/account/ifs/TokenHelperIfs.java
@@ -1,6 +1,5 @@
 package account.domain.account.ifs;
 
-import account.domain.account.model.TokenDto;
 import java.util.Map;
 
 public interface TokenHelperIfs {
