+++ board/src/main/java/board/common/config/web/WebConfig.java
@@ -10,8 +10,8 @@ import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import board.domain.user.interceptor.AuthorizationInterceptor;
-import board.domain.user.resolver.UserSessionResolver;
+import board.common.interceptor.AuthorizationInterceptor;
+import board.common.resolver.UserSessionResolver;
 
 @Component
 @RequiredArgsConstructor
