+++ board/src/main/java/board/common/interceptor/AuthorizationInterceptor.java
@@ -1,5 +1,6 @@
-package user.domain.user.interceptor;
+package board.common.interceptor;
 
+import board.common.exception.user.UserNotFoundException;
 import global.errorcode.UserErrorCode;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -11,7 +12,6 @@ import org.springframework.web.context.request.RequestAttributes;
 import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.servlet.HandlerInterceptor;
 import org.springframework.web.servlet.resource.ResourceHttpRequestHandler;
-import user.common.exception.user.UserNotFoundException;
 
 @Component
 @Slf4j
