+++ image/src/main/java/image/common/resolver/UserSessionResolver.java
@@ -1,7 +1,8 @@
-package pet.domain.user.resolver;
+package image.common.resolver;
 
 import db.domain.user.enums.UserRole;
 import global.annotation.UserSession;
+import image.domain.user.controller.model.User;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -13,7 +14,6 @@ import org.springframework.web.context.request.RequestAttributes;
 import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import pet.domain.user.controller.model.User;
 
 @Component
 @RequiredArgsConstructor
