+++ user/src/main/java/user/common/interceptor/AuthorizationInterceptor.java
@@ -1,6 +1,5 @@
-package board.domain.user.interceptor;
+package user.common.interceptor;
 
-import board.common.exception.user.UserNotFoundException;
 import global.errorcode.UserErrorCode;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -12,6 +11,7 @@ import org.springframework.web.context.request.RequestAttributes;
 import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.servlet.HandlerInterceptor;
 import org.springframework.web.servlet.resource.ResourceHttpRequestHandler;
+import user.common.exception.user.UserNotFoundException;
 
 @Component
 @Slf4j
